# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MatchHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}/groups:
    get:
      operationId: listUserGroups
      tags: [users]
      summary: List the private ranks a user belongs to with their position
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User group positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserGroupPosition"

  # Groups (private ranks)
  /groups:
    get:
      operationId: listMyGroups
      tags: [groups]
      summary: List private ranks the current user belongs to
      responses:
        "200":
          description: Groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a private rank
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/join:
    post:
      operationId: joinGroup
      tags: [groups]
      summary: Join a private rank by code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinGroupInput"
      responses:
        "200":
          description: Joined
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Private rank not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get a private rank
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "403":
          description: Not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/rankings:
    get:
      operationId: getGroupRankings
      tags: [groups]
      summary: Get FIFA and PES rankings for a private rank
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group rankings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupRankings"
        "403":
          description: Not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/matches:
    get:
      operationId: getGroupMatches
      tags: [groups]
      summary: List the matches that count toward a private rank
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Matches in this private rank
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
        "403":
          description: Not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /groups/{groupId}/leave:
    post:
      operationId: leaveGroup
      tags: [groups]
      summary: Leave a private rank
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Left
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /groups/{groupId}/end:
    post:
      operationId: endGroup
      tags: [groups]
      summary: End a private rank (creator only)
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Creator only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Parties
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    User:
      type: object
      required: [id, username, displayName, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
...
    Match:
      type: object
      required: [id, partyId, game, matchFormat, status, createdBy, createdAt, players]
      properties:
        id:
          type: integer
        partyId:
          type: integer
        game:
          type: string
          enum: [fifa, pes]
        matchFormat:
          type: string
          enum: [1v1, 2v2, 3v3]
        teamAScore:
          type: ["integer", "null"]
        teamBScore:
          type: ["integer", "null"]
        winnerTeam:
          type: ["string", "null"]
        winType:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed]
        createdBy:
          type: integer
        startedAt:
          type: ["string", "null"]
        finishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        players:
          type: array
          items:
            $ref: "#/components/schemas/MatchPlayer"
        creator:
          $ref: "#/components/schemas/User"
    MatchPlayer:
      type: object
      required: [id, matchId, userId, team, isSpectator, points, goalsFor, goalsAgainst]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        userId:
          type: integer
        team:
          type: ["string", "null"]
          enum: [team_a, team_b, null]
        isSpectator:
          type: boolean
        result:
          type: ["string", "null"]
          enum: [win, loss, null]
        points:
          type: integer
        goalsFor:
          type: integer
        goalsAgainst:
          type: integer
        user:
          $ref: "#/components/schemas/User"
    RankingEntry:
      type: object
      required: [rank, userId, username, displayName, points, matches, wins, losses, goalsFor, goalsAgainst, goalDifference, winRate]
      properties:
        rank:
          type: integer
        userId:
          type: integer
        username:
          type: string
        displayName:
          type: string
        points:
          type: integer
        matches:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        goalsFor:
          type: integer
        goalsAgainst:
          type: integer
        goalDifference:
          type: integer
        winRate:
          type: number
    GroupInput:
      type: object
      required: [name, memberIds]
      properties:
        name:
          type: string
        memberIds:
          type: array
          items:
            type: integer
    JoinGroupInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
...
    Group:
      type: object
      required: [id, name, createdBy, status, createdAt, members]
      properties:
        id:
          type: integer
        name:
          type: string
        createdBy:
          type: integer
        code:
          type: string
          description: Present only for the private rank creator
        status:
          type: string
          enum: [active, ended]
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
    GroupRankings:
      type: object
      required: [fifa, pes]
      properties:
        fifa:
          type: array
          items:
            $ref: "#/components/schemas/RankingEntry"
        pes:
          type: array
          items:
            $ref: "#/components/schemas/RankingEntry"
    UserGroupPosition:
      type: object
      required: [id, name, memberCount, position]
      properties:
        id:
          type: integer
        name:
          type: string
        memberCount:
          type: integer
        position:
          type: integer
...